backend/pyproject.toml
[tool.poetry]
name = "slack-integration"
version = "0.1.0"
description = "Slack integration service for Goosebump platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
crewai = "0.201.1"
crewai-tools = "0.75.0"
supabase = "^2.3.1"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
boto3 = "^1.34.0"
nest-asyncio = "^1.6.0"
nltk = "^3.8.1"
extruct = "0.13.0"
w3lib = "^2.1.2"
lxml = "4.9.3"
selenium = "^4.17.2"
webdriver-manager = "^4.0.1"
python-multipart = "^0.0.20"
google-auth-oauthlib = "^1.2.2"
google-api-python-client = "^2.172.0"
httpx = "^0.27.0"
email-validator = "^2.2.0"
playwright = "^1.40.0"
slack-sdk = "^3.27.0"
slack-bolt = "^1.18.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
redis = "^5.0.1"
celery = "^5.3.4"
requests = "^2.31.0"
aiofiles = "^23.2.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
cryptography = "^43.0.1"
beautifulsoup4 = "^4.12.0"
aiohttp = "^3.9.0"
trafilatura = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0" 
pytest-asyncio = "^0.23.5"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pre-commit = "^3.6.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]
